feat(auth): Implement MongoRefreshTokenRepository using RefreshTokenEntity
