+++ apps/backend/users-service/src/auth/infrastructure/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './mongo-refresh-token.repository';
